+++ app/src/main/java/com/example/timelytrack/activity/MainActivity.kt
@@ -5,14 +5,8 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.rememberNavController
-import com.example.timelytrack.TimelyTrackApplication
 import com.example.timelytrack.ui.components.BottomNavBar
-import com.example.timelytrack.viewmodel.LogEntryViewModelFactory
-import com.example.timelytrack.viewmodel.LogViewModel2
-
-//import com.example.timelytrack.viewmodel.LogViewModel
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -20,10 +14,7 @@ class MainActivity : ComponentActivity() {
         setContent {
             MaterialTheme {
                 Surface {
-//                    val logViewModel: LogViewModel2 = viewModel(
-//                        factory = LogEntryViewModelFactory((application as TimelyTrackApplication).repository)
-//                    )
-//                    val logViewModel: LogViewModel2 = viewModel()
+                    // Set up navigation logic here
                     val navController = rememberNavController()
                     BottomNavBar(navController = navController)
                 }
