+++ app/src/main/java/com/example/timelytrack/data/LogEntryContainer.kt
@@ -1,9 +1,10 @@
 package com.example.timelytrack.data
 
-import com.example.timelytrack.data.LogEntryDatabase
 import android.content.Context
 
 class LogEntryContainer(private val context: Context) {
+    // Simple form of dependency injection
+    // Single point of access for repo, ensures only one instance of repo is created
     val logEntryRepository: LogEntryRepository by lazy {
         LogEntryRepository(LogEntryDatabase.getLogEntryDatabase(context).logEntryDao())
     }
