+++ app/src/main/java/com/example/timelytrack/model/LogEntry.kt
@@ -1,10 +1,8 @@
 package com.example.timelytrack.model
 
-import android.media.metrics.LogSessionId
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import java.util.UUID
 
 // Primary information for the app
 // Need to implement custom fields down the road
@@ -14,7 +12,6 @@ data class LogEntry(
     @PrimaryKey (autoGenerate = true)
     val id: Int = 0,
 //    val id: String = UUID.randomUUID().toString(), // UUID is a unique identifier for each log entry
-
 
     @ColumnInfo(name = "category_id")
     val categoryId: String, // Reference to the category to which this log entry belongs
