+++ app/src/main/java/com/example/timelytrack/ui/components/BottomNavBar.kt
@@ -23,8 +23,6 @@ import androidx.navigation.compose.rememberNavController
 import com.example.timelytrack.ui.history.HistoryScreen
 import com.example.timelytrack.ui.home.HomeScreen
 import com.example.timelytrack.ui.profile.ProfileScreen
-//import com.example.timelytrack.viewmodel.LogViewModel
-import com.example.timelytrack.viewmodel.LogViewModel2
 
 @Preview(showBackground = true)
 @Composable
@@ -36,9 +34,8 @@ fun BottomNavigationBarPreview() {
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
-//fun BottomNavBar(logViewModel: LogViewModel, navController: NavHostController = rememberNavController()
-fun BottomNavBar(navController: NavHostController = rememberNavController()
-) {
+fun BottomNavBar(navController: NavHostController = rememberNavController()) {
+    // Layouting for the bottom navigation bar
     Scaffold(
         bottomBar = {
             BottomBarComponent(navController = navController)
@@ -57,6 +54,7 @@ fun BottomNavBar(navController: NavHostController = rememberNavController()
 
 @Composable
 fun BottomBarComponent(navController: NavHostController) {
+    // UI directives for NavigationBarItem
     NavigationBar {
         val navBackStackEntry by navController.currentBackStackEntryAsState()
         val currentRoute = navBackStackEntry?.destination?.route
@@ -70,8 +68,8 @@ fun BottomBarComponent(navController: NavHostController) {
                     launchSingleTop = true
                     restoreState = true
                 }
-            })
-
+            }
+        )
 
         NavigationBarItem(
             icon = { Icon(Icons.Filled.History, contentDescription = "History") },
@@ -84,7 +82,6 @@ fun BottomBarComponent(navController: NavHostController) {
                 }
             }
         )
-
         NavigationBarItem(
             icon = { Icon(Icons.Filled.Person, contentDescription = "Profile") },
             label = { Text("Profile") },
@@ -96,6 +93,5 @@ fun BottomBarComponent(navController: NavHostController) {
                 }
             }
         )
-        // ... (rest of your NavigationBarItems)
     }
 }
