+++ app/src/main/java/com/example/timelytrack/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
 package com.example.timelytrack.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
