--- app/src/main/java/com/example/timelytrack/viewmodel/LogViewModel2.kt
@@ -1,135 +1,0 @@
-package com.example.timelytrack.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.ViewModelProvider.AndroidViewModelFactory.Companion.APPLICATION_KEY
-import androidx.lifecycle.viewModelScope
-import androidx.lifecycle.viewmodel.initializer
-import androidx.lifecycle.viewmodel.viewModelFactory
-import com.example.timelytrack.TimelyTrackApplication
-import com.example.timelytrack.model.LogEntry
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.launch
-import com.example.timelytrack.data.LogEntryRepository
-import kotlinx.coroutines.flow.*
-import java.util.UUID
-
-
-class LogViewModel2(private val repository: LogEntryRepository) : ViewModel() {
-
-    // variables
-//    private val _logEntries = MutableStateFlow<List<LogEntry>>(emptyList()) //App will modify Private variable
-//    val logEntries: StateFlow<List<LogEntry>> = _logEntries.asStateFlow() // expose public variable and functions
-
-    val allLogEntries: StateFlow<List<LogEntry>> = repository.allItems
-
-
-    companion object {
-        val Factory: ViewModelProvider.Factory = viewModelFactory {
-            initializer {
-                val application = (this[APPLICATION_KEY] as TimelyTrackApplication)
-                LogViewModel2(application.container.logEntryRepository)
-            }
-        }
-    }
-
-
-
-
-    // Function to add a new log entry with the current start timestamp
-    fun addLogEntry2(categoryId: String) {
-        val newLogEntry = LogEntry(
-            categoryId = categoryId,
-            startTimestamp = System.currentTimeMillis()
-        )
-        viewModelScope.launch {
-            repository.insertLogEntry(newLogEntry)
-        }
-    }
-
-    // Function to delete a specific log entry by its ID
-    fun removeLogEntry2(logEntry: LogEntry) {
-        viewModelScope.launch {
-            repository.deleteLogEntry(logEntry)
-        }
-    }
-
-    // Function to complete the latest log entry by setting the end timestamp
-    fun completeLastLogEntry2() {
-        viewModelScope.launch {
-            // Get the latest log entry from the database
-            val latestEntry = repository.getAllLogEntries().firstOrNull()?.lastOrNull()
-            latestEntry?.let { entry ->
-                // Update the entry's end timestamp to the current time
-                entry.endTimestamp = System.currentTimeMillis()
-                repository.updateLogEntry(entry)
-            }
-        }
-    }
-
-
-
-}
-
-
-class LogEntryViewModelFactory(private val repository: LogEntryRepository) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(LogViewModel2::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return LogViewModel2(repository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
-
-
-
-
-//    fun addLogEntry() {
-//        // Create a new log entry and add it to the list
-//
-//        val newEntry = LogEntry(categoryId = "1", startTimestamp = System.currentTimeMillis())
-//        _logEntries.update { currentEntries -> currentEntries  + newEntry }
-//    }
-//
-//    fun completeLastLogEntry() {
-//        // Find the last log entry and update its endTimestamp
-//        // this is used by the Checkmark button
-//
-//        _logEntries.value.lastOrNull()?.let { lastLog ->
-//            val updatedLog = lastLog.copy(
-//                endTimestamp = System.currentTimeMillis()
-//            )
-//            _logEntries.update { currentEntries ->
-//                currentEntries.dropLast(1) + updatedLog
-//            }
-//        }
-//    }
-//
-//    fun removeLogEntry(logEntry: LogEntry) {
-//        // Remove a log entry from the list
-//        _logEntries.update { currentEntries -> currentEntries.filterNot { it == logEntry } }
-//    }
-//
-//
-//    fun clearLogEntries() {
-//        // Clear all log entries from the list
-//        _logEntries.value = emptyList()
-//    }
-
-//    //=== Retrieves a log entry by ID
-//    fun getLogEntryById(id: String): LogEntry? {
-//        return _logEntries.value.find { it.id == id }
-//    }
-
-//
-//    init {
-//        // Collect entries from the database to keep _logEntries updated
-//        viewModelScope.launch {
-//            repository.getAllLogEntries().collect { entries ->
-//                _logEntries.value = entries
-//            }
-//        }
-//    }
